Refactor: Improve image caching and use expo-image for ImageBackground
This commit improves image caching by changing the `cachePolicy` of the `expo-image` component from `disk` to `memory-disk`. This allows images to be cached in memory as well as on disk, which can improve performance.

Additionally, this commit changes the import of `ImageBackground` from `react-native` to `expo-image` in the following components:

- `components/Player.tsx`
- `components/SubPlayer.tsx`

This ensures that the `ImageBackground` component is using the same caching policy as the other images in the application.